@@ -1,10 +1,10 @@
 package quickxorhash
 
 import (
+	"crypto/rand"
 	"encoding/base64"
 	"fmt"
 	"hash"
-	"math/rand"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -171,7 +171,9 @@ var _ hash.Hash = (*quickXorHash)(nil)
 func BenchmarkQuickXorHash(b *testing.B) {
 	b.SetBytes(1 << 20)
 	buf := make([]byte, 1<<20)
-	rand.Read(buf)
+	n, err := rand.Read(buf)
+	require.NoError(b, err)
+	require.Equal(b, len(buf), n)
 	h := New()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
